idlib: make the memory statistics report real numbers
com_showMemoryUsage has been printing zero. USE_LIBC_MALLOC is 1, so mem_heap
is never created and Mem_Alloc returns local_malloc() before reaching
Mem_UpdateAllocStats; Mem_Free calls local_free() without recording anything.
Both counters therefore stayed at their initial values for the whole session,
and the registered memoryDump command was an empty stub, which left the engine
with no working way to say how much memory it was holding.

Recover the block size from the allocator rather than threading a size header
through every allocation, so the statistics cannot change the layout of what
they measure: malloc_size, malloc_usable_size or _msize depending on platform.
Record on all four libc paths, and on the heap path for Mem_Alloc16/Mem_Free16,
which were missing their accounting too.

The counters are written from every thread that allocates, so they are now
atomics updated with relaxed ordering instead of plain ints being raced on.
totalSize widens to int64_t: a loaded map measured 1.72 GB on an Android
device, within a few hundred megabytes of overflowing the old int.

Fill in memoryDump with the totals. The overlay cannot be captured in a log,
and a console command can.

Also fix idImage::StorageSize, which counted one face of a cube map instead of
six and ignored the sample count of a multisampled target, so listImages
under-reported reflection probes and the lightgrid.

Measured on game/convoy1: 169694 live allocations, 367.4 MB, and 314.7 MB of
images, where both figures previously read zero.

+++ src/idlib/Heap.h
@@ -117,7 +117,10 @@ typedef struct {
 	int		num;
 	int		minSize;
 	int		maxSize;
-	int		totalSize;
+	// A single allocation is bounded by Mem_ValidateAllocSize, but the running
+	// total is not: a loaded Quake 4 map sits within a few hundred megabytes of
+	// INT_MAX, so the accumulator has to be wider than the sizes it sums.
+	int64_t	totalSize;
 } memoryStats_t;
 
 // RAVEN BEGIN
